+++ src/cpu/core_full/save.h
@@ -117,5 +117,5 @@ switch (inst.code.save) {
 	case 0:
 		break;
 	default:
-		LOG(LOG_CPU,LOG_ERROR)("SAVE:Unhandled code %d entry %X",inst.code.save,inst.entry);
+		LOG(LOG_CPU,LOG_ERROR)("SAVE:Unhandled code %d entry %X",inst.code.save,static_cast<uint32_t>(inst.entry));
 }
